+++ src/Makevars
@@ -1,6 +1,6 @@
 OBJECTS= lib/ParquetReader.o RParquetReader.o RParquetAppender.o \
-  rwrapper.o protect.o read2.o write.o \
-  read-metadata.o read-pages.o \
+  RParquetOutFile.o r-utils.o rwrapper.o protect.o read2.o write2.o \
+  cpp-utils.o read-metadata.o read-pages.o \
   arrow-schema.o base64.o r-base64.o snappy.o encodings.o \
   dictionary-encoding.o test.o \
   lib/ParquetOutFile.o lib/RleBpDecoder.o \
