+++ src/read-metadata.cpp
@@ -12,8 +12,7 @@ extern "C" {
 
 extern SEXP nanoparquet_call;
 
-// Does not throw C++ exceptions, so we can wrap it
-SEXP convert_logical_type_(parquet::LogicalType ltype) {
+SEXP convert_logical_type(parquet::LogicalType ltype) noexcept {
   SEXP rtype = R_NilValue;
   int prot = 0;
   if (ltype.__isset.STRING) {
@@ -119,21 +118,6 @@ SEXP convert_logical_type_(parquet::LogicalType ltype) {
   return rtype;
 }
 
-SEXP convert_logical_type_wrapper(void *data) {
-  parquet::LogicalType *ltype = (parquet::LogicalType*) data;
-  return convert_logical_type_(*ltype);
-}
-
-SEXP convert_logical_type(parquet::LogicalType ltype, SEXP *uwt) {
-  return R_UnwindProtect(
-    convert_logical_type_wrapper,
-    &ltype,
-    throw_error,
-    uwt,
-    *uwt
-  );
-}
-
 SEXP convert_key_value_metadata(const parquet::FileMetaData &fmd) {
   SEXP uwtoken = PROTECT(R_MakeUnwindCont());
   R_API_START(R_NilValue);
@@ -217,7 +201,7 @@ SEXP convert_schema(const char *cfile_name,
     INTEGER(converted_type)[idx] =
     sch.__isset.converted_type ? sch.converted_type : NA_INTEGER;
     if (sch.__isset.logicalType) {
-      SET_VECTOR_ELT(logical_type, idx, convert_logical_type(sch.logicalType, &uwtoken));
+      SET_VECTOR_ELT(logical_type, idx, convert_logical_type(sch.logicalType));
     }
     INTEGER(num_children)[idx] =
       sch.__isset.num_children ? sch.num_children : NA_INTEGER;
